+++ pysys/utils/safeeval.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-# PySys System Test Framework, Copyright (C) 2006-2020 M.B.Grieve
+# PySys System Test Framework, Copyright (C) 2006-2021 M.B.Grieve
 
 # This library is free software; you can redistribute it and/or
 # modify it under the terms of the GNU Lesser General Public
@@ -39,7 +39,7 @@ import pysys
 import importlib
 from importlib import import_module
 
-def safe_eval(expr, extraNamespace={}):
+def safeEval(expr, extraNamespace={}, errorMessage='Failed to evaluate "{expr}" due to {error}'):
 	"""
 	Executes eval(...) on the specified string expression, using a controlled globals()/locals() environment to 
 	ensure we do not break compatibility between PySys versions, and that a sensible set of PySys constants and 
@@ -48,18 +48,30 @@ def safe_eval(expr, extraNamespace={}):
 	The global environment used for evaluation includes the ``os.path``, ``math``, ``sys``, ``re``, ``json``, and ``locale`` 
 	standard Python modules, as well as the ``pysys`` module and the contents of the `pysys.constants` module, e.g. ``IS_WINDOWS``. 
 	
-	If necessary, symbols for additional modules can be imported dynamically using ``import_module``, e.g. 
-	``safe_eval("import_module('difflib').get_close_matches('app', ['apple', 'orange', 'applic']")``.
+	If necessary, symbols for additional modules can be imported dynamically using ``import_module``:: 
 	
-	:param expr: The string to be evaluated.
+		x = safeEval("import_module('difflib').get_close_matches('app', ['apple', 'orange', 'applic']")
 	
-	:param extraNamespace: A dict of string names and Python object values to be included in the globals dict 
+	If an error occurs, an exception is raised that includes the expression in its message. 
+	
+	:param str expr: The string to be evaluated.
+	
+	:param dict[str,obj] extraNamespace: A dict of string names and Python object values to be included in the globals dict 
 		used to evaluate this string. 
+	
+	:param str errorMessage: The string used for the raised exception message if an exception is thrown by eval, 
+		where ``{expr}`` will be replaced with the actual expression and ``{error}`` with the error message. 
+		
+	.. versionadded:: 1.7.0
 	"""
 	env = globals()
 	if extraNamespace:
 		env = dict(env)
 		for k,v in extraNamespace.items():
 			env[k] = v
-	return eval(expr, env)
+	try:
+		return eval(expr, env)
+	except Exception as e:
+		raise Exception(errorMessage.format(expr=expr.replace('\n',' ').replace('\r','').strip(), error='%s - %s'%(e.__class__.__name__, e or '<no message>')).strip())
+