+++ samples/getting-started/test/correctness/MyServer_003/run.py
@@ -2,20 +2,20 @@ import pysys
 from pysys.constants import *
 
 class PySysTest(pysys.basetest.BaseTest):
+
+	compressionType = None 
+	""" This test can be run in multiple modes. The self.mode string indicates which one we're executing, but 
+	usually it's easier to use the fact that any parameters defined on the mode are set on the test object 
+	(and also available as self.mode.params). So compressionType will be set on the test object based on which 
+	mode we're running under. """
+	
+	auth = None
+
 	def execute(self):
 		server = self.myserver.startServer(name="my_server")
 		
-		# This test can be run in multiple modes. The self.mode string indicates which one we're executing. 
-		# In this case we have multiple dimensions to the mode so we need to unpack them. This could also be done in 
-		# a test plugin class if these modes were used by multiple tests. 
-		compressionType, auth = self.mode.split('_')
-		compressionType = {
-			'CompressionGZip': 'gzip', 
-			'CompressionNone': '',
-			}[compressionType]
-		
 		self.startPython([self.input+'/httpget.py', 
-			f'http://127.0.0.1:{server.info["port"]}/sensorValues', compressionType, auth], stdouterr='sensorValues')
+			f'http://127.0.0.1:{server.info["port"]}/sensorValues', self.compressionType, self.auth], stdouterr='sensorValues')
 
 	def validate(self):	
 		self.logFileContents('sensorValues.out', maxLines=0)